// lib/data/enums.ts
export enum Degree {
  Min = 0,
  Max = 360,
}

export enum PercentRange {
  Min = 0,
  Max = 100,
}

export enum UsernameLength {
  Min = 1,
  Max = 30,
}

export enum CardBorderRadius {
  Min = 0,
  Max = 50,
}

export enum CardBorderWidth {
  Min = 0,
  Max = 10,
}

export enum LayoutMaxWidth {
  Min = 400,
  Max = 1000,
}

export enum ContainerBorderWidth {
  Min = 0,
  Max = 5,
}

export enum PageTransitionDuration {
  Min = 0.2,
  Max = 3,
}